refactor: integra Spring Data JPA com PostgreSQL

+++ src/main/java/exception/EmprestimoNaoEncontradoException.java
@@ -1,7 +1,7 @@
 package exception;
 
 public class EmprestimoNaoEncontradoException extends BibliotecaException {
-    public EmprestimoNaoEncontradoException(int id) {
+    public EmprestimoNaoEncontradoException(Long id) {
         super("Empréstimo não encontrado com id: " + id);
     }
 }
