+++ src/main/java/exception/LivroNaoEncontradoException.java
@@ -1,7 +1,7 @@
 package exception;
 
 public class LivroNaoEncontradoException extends BibliotecaException {
-  public LivroNaoEncontradoException(int id) {
+  public LivroNaoEncontradoException(Long id) {
     super("Livro não encontrado com id: " + id);
   }
 }
