+++ src/main/java/exception/UsuarioNaoEncontradoException.java
@@ -1,7 +1,7 @@
 package exception;
 
 public class UsuarioNaoEncontradoException extends BibliotecaException {
-  public UsuarioNaoEncontradoException(int id) {
+  public UsuarioNaoEncontradoException(Long id) {
     super("Usuário não encontrado com id: " + id);
   }
 }
